ConfirmActivity added
